requirements.txt
oandapyV20>=0.7.2
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
plotly>=5.18.0
jinja2>=3.1.0
schedule>=1.2.0
colorama>=0.4.6
jsonlines>=4.0.0
tabulate>=0.9.0
requests>=2.28.0

